Fixed first exception from SVY-3048 : Deleting forms in Servoy Developer leave some resource references which the build process later attempts erroneously to use.
It's actually about references to closed or deleted resources projects.
Case: SVY-3048

+++ com.servoy.eclipse.model/src/com/servoy/eclipse/model/builder/ServoyBuilder.java
@@ -3765,7 +3765,7 @@ public class ServoyBuilder extends IncrementalProjectBuilder
 			{
 				for (IProject referenced : project.getReferencedProjects())
 				{
-					if (referenced.hasNature(ServoyResourcesProject.NATURE_ID))
+					if (referenced.exists() && referenced.isOpen() && referenced.hasNature(ServoyResourcesProject.NATURE_ID))
 					{
 						deleteMarkers(referenced, COLUMN_MARKER_TYPE);
 					}
